requirements.txt
tensorflow
tensorflow-cpu
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart